lint
SCCS-vsn: sys/vax/mba/ht.c 4.33
SCCS-vsn: sys/kern/kern_descrip.c 5.14
SCCS-vsn: sys/kern/uipc_syscalls.c 4.33
SCCS-vsn: sys/ufs/ffs/dinode.h 4.16
SCCS-vsn: sys/ufs/ffs/inode.h 4.16
SCCS-vsn: sys/ufs/ufs/dinode.h 4.16
SCCS-vsn: sys/ufs/ufs/inode.h 4.16
SCCS-vsn: sys/sys/kernel.h 4.5
SCCS-vsn: sys/ufs/ffs/quota.h 4.2
SCCS-vsn: sys/ufs/ufs/quota.h 4.2
SCCS-vsn: sys/sys/systm.h 4.29
SCCS-vsn: sys/sys/types.h 4.7
SCCS-vsn: sys/net/if.c 4.22
SCCS-vsn: sys/net/raw_usrreq.c 4.21
SCCS-vsn: sys/netinet/in.c 4.7
SCCS-vsn: sys/netinet/in.h 4.16
SCCS-vsn: sys/netinet/in_pcb.c 4.34
SCCS-vsn: sys/netinet/in_pcb.h 4.5
SCCS-vsn: sys/netinet/ip_icmp.c 4.22
SCCS-vsn: sys/netinet/ip_input.c 1.54
SCCS-vsn: sys/netinet/ip_output.c 1.40
SCCS-vsn: sys/netinet/tcp_input.c 1.80
SCCS-vsn: sys/netinet/tcp_timer.c 4.28
SCCS-vsn: sys/netinet/tcp_usrreq.c 1.68
SCCS-vsn: sys/netinet/udp_usrreq.c 4.37
SCCS-vsn: sys/vax/vax/conf.c 4.58
SCCS-vsn: sys/vax/vax/machdep.c 4.65
SCCS-vsn: sys/vax/vax/mem.c 4.8
SCCS-vsn: sys/vax/vax/trap.c 4.20
SCCS-vsn: sys/vax/vax/tu.c 4.11
SCCS-vsn: sys/vax/uba/dmf.c 4.12
SCCS-vsn: sys/vax/uba/ts.c 4.35
SCCS-vsn: sys/vax/uba/uba.c 4.53
SCCS-vsn: sys/deprecated/netimp/if_imp.c 4.41

+++ usr/src/sys/vax/vax/mem.c
@@ -1,4 +1,4 @@
-/*	mem.c	4.7	82/10/17	*/
+/*	mem.c	4.8	82/10/20	*/
 
 /*
  * Memory special file
@@ -62,8 +62,8 @@ mmrw(dev, uio, rw)
 			*(int *)mmap = v | (PG_V | PG_KR);
 			mtpr(TBIS, vmmap);
 			o = (int)uio->uio_offset & PGOFSET;
-			c = min((unsigned)(NBPG - o), iov->iov_len);
-			c = min(c, (unsigned)(NBPG - ((int)iov->iov_base&PGOFSET)));
+			c = min((u_int)(NBPG - o), (u_int)iov->iov_len);
+			c = min(c, (u_int)(NBPG - ((int)iov->iov_base&PGOFSET)));
 			error = uiomove((caddr_t)&vmmap[o], c, rw, uio);
 			continue;
 
@@ -96,7 +96,7 @@ mmrw(dev, uio, rw)
 			if (!useracc(iov->iov_base, c, rw == UIO_READ ? B_WRITE : B_READ))
 				goto fault;
 			error = UNIcpy((caddr_t)uio->uio_offset, iov->iov_base,
-			    c, rw);
+			    (int)c, rw);
 			break;
 		}
 		if (error)
@@ -116,7 +116,7 @@ fault:
  */
 UNIcpy(uniadd, usradd, n, rw)
 	caddr_t uniadd, usradd;
-	register u_int n;
+	register int n;
 	enum uio_rw rw;
 {
 	register short *from, *to;
